requirements.txt
langchain==0.3.25
langchain-groq==0.3.2
langchain-community==0.3.24
python-dotenv==1.0.1
flask==3.1.0
